requirements.txt

mysql-connector-python
prettytable
